requirements.txt
paho-mqtt
psutil
dash
dash_bootstrap_components